+++ Product/src/main/java/services/api/IProductService.java
@@ -3,5 +3,5 @@ package services.api;
 import entity.core.Product;
 
 public interface IProductService extends IEssenceService<Product> {
-	 void addNewProduct(int id, String title, int price, int discount, String description);
+	void addNewProduct(int id, String title, double price, double discount, String description);
 }
